Automatically backed up by Learn

+++ index.js
@@ -30,7 +30,7 @@ function makeInt(n) {
 }
 
 function preserveDecimal(n) {
- return parseFloat(n, 10)
+ return parseInt(n, 10)
 }
 
 function preserveDecimal(n) {
